Extract drillPractice | saveBurglarState | burglarDefense | bonusSass
